backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
supabase==2.7.4
httpx==0.27.2
google-generativeai==0.8.3
elevenlabs==1.9.0
reportlab==4.2.2
python-multipart==0.0.12
pydantic==2.9.2
